Move a Work's author property to its Manifestations
A Work is the intangible representation of a Manifestation, so the
author field does not necessarily apply. The holder of the Copyright is
a better representation of who the "owner" of the Work is; the "creator"
property should be used instead to denote the original creator (if not
already implicitly specified by the Copyright).